requirements-prod.txt
# Production Requirements for Vercel Deployment
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
WTForms==3.0.1
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0

# Fix protobuf version conflict
protobuf<6.0.0

# AI and Machine Learning
google-generativeai>=0.3.0
scikit-learn==1.4.1.post1
spacy==3.7.2
numpy==1.26.4
pandas==2.2.1

# Schedule Optimization
ortools>=9.8.0

# Calendar Integration
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
httplib2>=0.22.0

# Health Data Processing
requests==2.31.0

# Additional utilities
python-dateutil==2.8.2
pytz==2024.1

# Additional dependencies for proper functionality
setuptools>=65.0.0
wheel>=0.37.0
cachetools>=4.2.0
urllib3>=1.26.0,<2.0.0
certifi>=2022.12.7

# Production server
gunicorn==21.2.0